third_party/WebKit/LayoutTests/inspector/coverage/decorations-after-script-formatter.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="../../http/tests/inspector/sources-test.js"></script>
<script src="coverage-test.js"></script>

<script src="resources/coverage.js"></script>

<script>
async function test()
{
    InspectorTest.startCoverage();
    await InspectorTest.evaluateInPagePromise("performActions()");
    await InspectorTest.stopCoverage();
    await UI.inspectorView.showPanel("sources");
    await InspectorTest.sourceDecorated("coverage.js");

    var decoratorPromise = InspectorTest.addSnifferPromise(Coverage.CoverageView.LineDecorator.prototype, "_innerDecorate");
    var editorActions = await self.runtime.allInstances(Sources.SourcesView.EditorAction);
    for (action of editorActions) {
        if (action instanceof Sources.ScriptFormatterEditorAction) {
            action._toggleFormatScriptSource();
            break;
        }
    }
    await decoratorPromise;
    InspectorTest.dumpDecorationsInSourceFrame(UI.panels.sources.visibleView);
    InspectorTest.completeTest();
}
</script>
</head>
<p id="id">PASS</p>
<body onload="runTest()">
<p>Tests the gutter decorations in target source code after ScriptFormatterEditorAction</p>
</body>
</html>
