third_party/WebKit/LayoutTests/inspector/coverage/decorations-after-script-formatter-expected.txt

PASS

Tests the gutter decorations in target source code after ScriptFormatterEditorAction

0: + function outer(index) {
1:   
2: -     function inner1(a) {
3: -         return a + 1;
4: -     }
5:   
6: +     function inner2(a) {
7: +         return a + 2;
8: +     }
9:   
10: -     function inner3(a) {
11: -         return a + 3;
12: -     }
13: +     function inner4(a) {
14: +         return a + 4;
15: +     }
16: -     function inner5(a) {
17: -         return a + 5;
18: -     }
19:   
20: +     return [inner1, inner2, inner3, inner4, inner5][index];
21: + }
22:   
23: + function performActions() {
24: +     return outer(1)(0) + outer(3)(0);
25: + }
26: - function outer2() {
27: -     return outer(0)(0);
28: - }
29:   
